Fix critical bug that was resulted to have duplicate items in graph
Changed graph structure to resolve based on normalized list of
dependencies, lockfile contains duplicates with a reference as there are
different intents in package.json files to resolve the same dependency.
The normalized lock file entry is intended to have list of resolved
entries which is the same for all inents.